+++ src/main/java/com/kakaotech/team14backend/point/appliation/UsePointUsecase.java
@@ -3,9 +3,9 @@ package com.kakaotech.team14backend.point.appliation;
 import com.kakaotech.team14backend.common.MessageCode;
 import com.kakaotech.team14backend.exception.MemberNotFoundException;
 import com.kakaotech.team14backend.point.domain.Point;
-import com.kakaotech.team14backend.inner.point.model.PointHistory;
-import com.kakaotech.team14backend.inner.point.model.TransactionType;
-import com.kakaotech.team14backend.inner.point.repository.PointHistoryRepository;
+import com.kakaotech.team14backend.point.domain.PointHistory;
+import com.kakaotech.team14backend.point.domain.TransactionType;
+import com.kakaotech.team14backend.post.infrastructure.PointHistoryRepository;
 import com.kakaotech.team14backend.point.infrastructure.PointRepository;
 import javax.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
